Port members over to its own component
